# [core][optimization] use a pool of numpy ndarray to hold seq data