@youkaichao 2024-06-27
following the setup in https://github.com/vllm-project/vllm/pull/5584#issuecomment-2172218557 , run them again:

script: `python benchmarks/benchmark_throughput.py --output-len 256 --dataset ShareGPT_V3_unfiltered_cleaned_split.json --model meta-llama/Llama-2-7b-hf -tp 1`

# Using a single H100 and meta-llama/Llama-2-7b-hf

before:

Throughput: 18.57 requests/s, 9530.05 tokens/s

After:

Throughput: 18.72 requests/s, 9607.86 tokens/s

# using H100 * 8 and meta-llama/Meta-Llama-3-70B

before: 

Throughput: 19.02 requests/s, 9134.99 tokens/s

after:

Throughput: 19.20 requests/s, 9220.82 tokens/s

# using H100 * 4 and meta-llama/Meta-Llama-3-70B

before: 

Throughput: 15.19 requests/s, 7295.74 tokens/s

after:

Throughput: 15.34 requests/s, 7366.30 tokens/s

# The following data is not updated. Don't have A100 at present.
# Using A100 * 4:

before:

~~Throughput: 16.84 requests/s, 8642.64 tokens/s~~

After:

~~Throughput: 16.58 requests/s, 8510.95 tokens/s~~

# Using a single A100:

before:

~~Throughput: 9.30 requests/s, 4773.91 tokens/s~~

after:

~~Throughput: 9.32 requests/s, 4786.32 tokens/s~~